refactor run.py, hdfs->pybhdfs
